Migrate from awoo-detat/awoo
This moves the bulk of the code from `awoo-detat/awoo` -- the web app
version of Werewolf we started on a few years ago -- to this repo, so it
can be used by the web app, a Discord bot (`awoo-detat/moon`) and, in
theory, other things if people feel like it.

There's work to be done to get back to the baseline functionality of
`awoo`, in particular around player messaging. I tossed some thoughts
into the `README` so I wouldn't forget.